Test and documenation CheckDeps
* CheckDeps_C added to ZmakeGcc.jzTc.sh
* src/test/ZmakeGcc/All_Test/applstdef_UseCCdef/applstdef_emC.h can be more individually, include path in cc_def variable

+++ +clean.bat
@@ -1,5 +1,4 @@
 @echo off
-call src\mklib\VS15-IDE\+clean.bat
-rmdir .gradle
-rmdir build
-pause
+if exist .gradle rmdir .gradle
+if exist build rmdir build
+if not "%1"=="nopause" pause
